Fixed KeyError in Font Size for HTML
Fixed an error in html.py where the font-size attribute from Box was not valid.  We were getting a value of 12 when the only known values are "medium" "large" "verylarge" and "small" as strings.  I wrapped the line of code where the KeyError was occuring in a try/except block, and if the exception is thrown, we will just pass (and keep the font size the same as it was for the previous element).